refactor: moving folders to src
