+++ src/helpers/token.js
